change uint type to uint64, fix some errors

+++ entity/model/entity.go
@@ -2,6 +2,6 @@ package model
 
 // BaseEntity 实体的基本接口，包含被框架托管的必须方法
 type BaseEntity interface {
-	TypeID() uint
-	InjectTypeID(id uint)
+	TypeID() uint64
+	InjectTypeID(id uint64)
 }
